requirements.txt
Django==3.2.13
drf-yasg==1.20.0
pre-commit==2.19.0
dj-rest-auth==2.2.4
django-environ==0.8.1
django-allauth==0.50.0
